requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Babel==2.15.0
backcall==0.2.0
bce-python-sdk==0.9.17
blinker==1.8.2
bokeh==3.1.1
boltons==24.0.0
Bottleneck==1.4.0
braceexpand==0.1.7
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.1.1
cryptography==43.0.0
cycler==0.12.1
Cython==3.0.10
datasets==2.20.0
decorator==5.1.1
dill==0.3.4
Distance==0.1.3
dnspython==2.6.1
editdistance==0.8.1
einops==0.8.0
email_validator==2.2.0
environs==9.5.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
Flask==3.0.3
flask-babel==4.0.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
future==1.0.0
g2p-en==2.1.0
g2pM==0.1.2.5
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.0.0
intervaltree==3.1.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==4.0.0
kaldiio==2.18.0
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.41.1
loguru==0.7.2
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mido==1.3.2
mock==5.1.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.12.2
nara-wpe==0.0.10
nltk==3.8.1
note-seq==0.0.3
numba==0.58.1
numpy==1.24.4
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.1
OpenCC==1.1.7
opencc-python-reimplemented==0.1.7
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==23.2
paddle-bfloat==0.1.7
paddle2onnx==1.0.6
paddleaudio==1.1.0
paddlefsl==1.1.0
paddlenlp==2.6.1
paddlepaddle==2.4.1
paddlesde==0.2.5
paddleslim==2.6.0
paddlespeech==1.4.1
paddlespeech-feat==0.1.0
pandas==2.0.3
parameterized==0.9.0
parso==0.8.4
pathos==0.2.8
pattern_singleton==1.2.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
portalocker==2.10.1
pox==0.3.4
ppdiffusers==0.19.4
ppft==1.7.6.8
praatio==5.1.1
pretty_midi==0.2.10
prettytable==3.10.2
prompt_toolkit==3.0.47
protobuf==3.20.2
psutil==6.0.0
pure-eval==0.2.2
pyarrow==17.0.0
pyarrow-hotfix==0.6
pybind11==2.13.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pymilvus==2.4.4
PyMySQL==1.1.1
pyparsing==3.1.2
pypinyin==0.44.0
pypinyin-dict==0.8.0
pyreadline3==3.4.1
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
pyworld==0.3.4
PyYAML==6.0.1
pyzmq==26.0.3
rarfile==4.2
regex==2024.5.15
requests==2.32.3
requests-mock==1.12.1
resampy==0.4.3
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.2.0
seqeval==1.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.3
starlette==0.37.2
swig==4.2.1
sympy==1.13.1
tabulate==0.9.0
TextGrid==1.6.1
threadpoolctl==3.5.0
timer==0.3.0
ToJyutping==0.3.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trampoline==0.1.2
typeguard==2.13.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==1.26.19
uvicorn==0.30.3
visualdl==2.5.3
watchfiles==0.22.0
wcwidth==0.2.13
webrtcvad==2.0.10
websockets==12.0
Werkzeug==3.0.3
win32-setctime==1.1.0
xxhash==3.4.1
xyzservices==2024.6.0
yacs==0.1.8
yarl==1.9.4
zhon==2.0.2
zipp==3.19.2
